Fixing the problems with AbstractModule#rename:
Renaming the file is a potential mem leak since AM is supposed to
reregister his file listeners when he changes the path of its
descriptor.
Considering the fact that the whole file system managing logic has been
extracted from AM into external entities (ModuleFileTracker),
this commit is about the extracting #rename functionality into workbench.

AM must not change his path during its lifetime since we need to track
those renames and properly manage the file system listening in the
classes outside of AM.
That leaves us with two choices: one is to extract the whole rename back
into workbench, and another one is to propagate some
rename/move/delete events to those external module fs managers which respond in a correct way.

I think that the former is the best solution, since 1) it allows us to
have immutable AM which does not change its location; whenever the AM is
moved it should be just a brand new AM. Thus we avoid the potential fs leaks all
at once, since we do not allow such actions.
2) still it is possible to simply change the name of the module by
AM#getModuleDescriptor#setModuleName()
3) ideologically, we have AM tending not to deal with FS directly rather
having a guy on the side which does that for him;
So changing the property name of AM is totally available but dealing
with FS directly (like renaming and moving folders) must be avoided.
4) the rename event itself smells bad because the level of
ModuleFileTracker is essentially FS level, and all it does is listening
to fs events, updating AMs accordingly.
I assert that FS level is lower
than the module level, and ModuleFileTracker also responding to some essentially
user actions (rename events) would be mixing of two different levels and thus is a poorer design.

+++ core/model/source/jetbrains/mps/model/ModelDeleteHelper.java
@@ -87,11 +87,15 @@ public class ModelDeleteHelper {
       // e.g. IdeCommandUtil is more careful when executing #removeGenSources console command
       if (modelOutput != null) {
         possiblyEmptyDirsToPrune.add(modelOutput.getParent());
-        modelOutput.delete();
+        if (modelOutput.exists()) {
+          modelOutput.delete();
+        }
       }
       if (modelCaches != null) {
         possiblyEmptyDirsToPrune.add(modelCaches.getParent());
-        modelCaches.delete();
+        if (modelCaches.exists()) {
+          modelCaches.delete();
+        }
       }
     }
